apps/api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.9.1
python-multipart==0.0.20
duckdb>=1.4.0,<1.5
pandas>=2.3.0,<2.4
openpyxl==3.1.5
PyYAML==6.0.2
sqlglot==26.21.0
claude-agent-sdk>=0.0.1,<1.0
anthropic>=0.40.0,<1.0
bcrypt>=4.0.0,<6.0
itsdangerous>=2.2.0,<3.0
